rust: rename safeinputstring.rs to input.rs
input::SafeInputstring is shorter and just as descriptive.
